修复云端协议数据包构建中的温度湿度精度问题
- 修改 cloud_protocol.c 中的数据包构建函数，将整数除法改为浮点数格式
- 添加浮点数支持以正确显示温度和湿度的小数值
- 更新 snprintf 格式字符串以支持小数点一位的浮点数输出
- 添加注释说明各参数含义和浮点数转换逻辑
- 修复 network_core.c 中的时间同步间隔宏定义计算方式

+++ User/cloud_protocol.c
@@ -15,15 +15,16 @@ uint16_t Cloud_Build_Upload_Packet(
     uint8_t dehumidifier, uint8_t humidifier,
     int16_t temp_x10, int16_t hum_x10)
 {
+    // ✅ 使用浮点数格式，支持小数和负数
     int n = snprintf(buffer, buf_size,
-        "cmd=2&uid=%s&topic=data&msg=Mode:%d th:%d tl:%d hh:%d hl:%d "
+        "cmd=2&uid=%s&topic=data&msg=Mode:%d th:%.1f tl:%.1f hh:%.1f hl:%.1f "
         "jr:%d zl:%d cs:%d js:%d temp:%d.%d humi:%d.%d\r\n",
         BEMFA_UID,
         mode,
-        temp_high / 10,
-        temp_low / 10,
-        humid_high / 10,
-        humid_low / 10,
+        (float)temp_high / 10.0f,   // th: 温度上限（浮点数，支持负数）
+        (float)temp_low / 10.0f,    // tl: 温度下限（浮点数，支持负数）
+        (float)humid_high / 10.0f,  // hh: 湿度上限（浮点数）
+        (float)humid_low / 10.0f,   // hl: 湿度下限（浮点数）
         heater, cooler, dehumidifier, humidifier,
         temp_x10 / 10,
         (temp_x10 < 0 ? -temp_x10 : temp_x10) % 10,
